requirements.txt
streamlit==1.35.0
tensorflow-cpu==2.12.0
pandas
numpy
scikit-learn
joblib
matplotlib
seaborn
